pzsp/backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
numpy
opencv-python
